Changed app to 'auth'
